requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy>=2.0.36
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==3.0.1
